+++ packages/core/plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@loglayer/plugin` Changelog
 
+## 2.2.4-alpha-0.0
+
+### Patch Changes
+
+- Updated dependencies [d5bb7f7]
+  - @loglayer/shared@3.1.0-alpha-0.0
+
 ## 2.2.3
 
 ### Patch Changes