+++ packages/core/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # `@loglayer/shared` Changelog
 
+## 3.1.0-alpha-0.0
+
+### Minor Changes
+
+- d5bb7f7: Better mixin support
+
 ## 3.0.1
 
 ### Patch Changes