+++ packages/plugins/datadog-apm-trace-injector/CHANGELOG.md
@@ -1,5 +1,13 @@
 # `@loglayer/datadog-apm-trace-injector` Changelog
 
+## 1.0.13-alpha-0.0
+
+### Patch Changes
+
+- Updated dependencies [d5bb7f7]
+  - @loglayer/shared@3.1.0-alpha-0.0
+  - @loglayer/plugin@2.2.4-alpha-0.0
+
 ## 1.0.12
 
 ### Patch Changes