+++ packages/transports/datadog/CHANGELOG.md
@@ -1,5 +1,11 @@
 # `@loglayer/transport-datadog` Changelog
 
+## 3.3.1-alpha-0.0
+
+### Patch Changes
+
+- @loglayer/transport@2.3.10-alpha-0.0
+
 ## 3.3.0
 
 ### Minor Changes