+++ packages/transports/log-file-rotation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # `@loglayer/transport-log-file-rotation` Changelog
 
+## 2.1.13-alpha-0.0
+
+### Patch Changes
+
+- @loglayer/transport@2.3.10-alpha-0.0
+
 ## 2.1.12
 
 ### Patch Changes