+++ packages/transports/pino/CHANGELOG.md
@@ -1,5 +1,11 @@
 # `@loglayer/transport-pino` Changelog
 
+## 2.2.14-alpha-0.0
+
+### Patch Changes
+
+- @loglayer/transport@2.3.10-alpha-0.0
+
 ## 2.2.13
 
 ### Patch Changes