+++ packages/transports/sentry/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @loglayer/transport-sentry
 
+## 1.0.6-alpha-0.0
+
+### Patch Changes
+
+- @loglayer/transport@2.3.10-alpha-0.0
+
 ## 1.0.5
 
 ### Patch Changes