+++ packages/transports/victoria-logs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.17-alpha-0.0
+
+### Patch Changes
+
+- @loglayer/transport-http@1.1.8-alpha-0.0
+
 ## 1.0.16
 
 ### Patch Changes